chore: post 디렉터리 구조 정리
