+++ src/main/java/core/domain/board/service/BoardService.java
@@ -1,7 +1,7 @@
 package core.domain.board.service;
 
 import core.domain.board.dto.CategoryListResponse;
-import core.domain.post.dto.PostWriteAnonymousAvailableResponse;
+import core.domain.post.dto.comunity.PostWriteAnonymousAvailableResponse;
 
 import java.util.List;
