+++ src/main/java/core/domain/board/service/impl/BoardServiceImpl.java
@@ -4,7 +4,7 @@ import core.domain.board.dto.CategoryListResponse;
 import core.domain.board.entity.Board;
 import core.domain.board.repository.BoardRepository;
 import core.domain.board.service.BoardService;
-import core.domain.post.dto.PostWriteAnonymousAvailableResponse;
+import core.domain.post.dto.comunity.PostWriteAnonymousAvailableResponse;
 import core.global.enums.BoardCategory;
 import core.global.exception.BusinessException;
 import core.global.enums.errorcode.CommunityErrorCode;
