+++ src/main/java/core/domain/post/controller/CrawlerController.java
@@ -1,8 +1,8 @@
 package core.domain.post.controller;
 
-import core.domain.post.service.KLifeCrawlerService;
-import core.domain.post.service.KoreaNetCrawlerService;
-import core.domain.post.service.SeoulGlobalCrawlerService;
+import core.domain.post.service.crawling.KLifeCrawlerService;
+import core.domain.post.service.crawling.KoreaNetCrawlerService;
+import core.domain.post.service.crawling.SeoulGlobalCrawlerService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
