+++ src/main/java/core/domain/post/controller/PostSearchController.java
@@ -1,12 +1,12 @@
 package core.domain.post.controller;
 
-import core.domain.post.dto.PostDetailResponse;
-import core.domain.post.dto.SearchResultView;
-import core.domain.post.dto.SuggestClickRequest;
+import core.domain.post.dto.comunity.PostDetailResponse;
+import core.domain.post.dto.search.SearchResultView;
+import core.domain.post.dto.search.SuggestClickRequest;
 import core.domain.post.service.PostSearchService;
 import core.domain.post.service.PostService;
-import core.domain.post.service.RecentSearchRedisService;
-import core.domain.post.service.SuggestMemoryIndex;
+import core.domain.post.service.search.RecentSearchRedisService;
+import core.domain.post.service.search.SuggestMemoryIndex;
 import core.global.dto.ApiResponse;
 import core.global.metrics.FeatureUsageMetrics;
 import core.global.pagination.CursorPageResponse;
