--- src/main/java/core/domain/post/dto/SearchResult.java
@@ -1,8 +1,0 @@
-package core.domain.post.dto;
-
-public record SearchResult(
-        long id,
-        String title,
-        String snippet,
-        double score
-) {}
