+++ src/main/java/core/domain/post/repository/impl/PostRepositoryImpl.java
@@ -11,10 +11,10 @@ import com.querydsl.jpa.impl.JPAQueryFactory;
 import core.domain.board.dto.BoardItem;
 import core.domain.board.entity.QBoard;
 import core.domain.comment.entity.QComment;
-import core.domain.post.dto.PostDetailResponse;
-import core.domain.post.dto.PostListResponse;
-import core.domain.post.dto.PostSearchRequest;
-import core.domain.post.dto.UserPostItem;
+import core.domain.post.dto.comunity.PostDetailResponse;
+import core.domain.post.dto.admin.PostListForAdminResponse;
+import core.domain.post.dto.admin.PostSearchForAdminRequest;
+import core.domain.post.dto.comunity.UserPostItem;
 import core.domain.post.entity.Post;
 import core.domain.post.entity.QBlockPost;
 import core.domain.post.entity.QPost;
@@ -411,57 +411,11 @@ public class PostRepositoryImpl implements PostRepositoryCustom {
                 .fetch();
     }
 
-    @Override
-    public List<BoardItem> findPostsByIdsForSearch(Long viewerId, List<Long> ids) {
-        if (ids == null || ids.isEmpty()) return List.of();
-
-        Expression<String> preview = preview200();
-        Expression<Long> likeCountExpr = likeCountExpr();
-        Expression<Long> commentCountExpr = commentCountExpr();
-        Expression<Boolean> likedByMe = likedByViewerId(viewerId);
-        Expression<Long> authorIdExpr = authorIdExpr();
-        BooleanExpression visibleToMe = visibleTo(viewerId);
-        Expression<Boolean> bookmarkedByMe = bookmarkedByViewerId(viewerId);
-
-        QImage uimg = new QImage("uimg");
-        Expression<String> userImageUrlExpr =
-                JPAExpressions.select(uimg.url)
-                        .from(uimg)
-                        .where(uimg.imageType.eq(IMAGE_TYPE_USER)
-                                .and(uimg.relatedId.eq(user.id)));
-
-        Expression<String> contentThumbUrlExpr = firstPostImageUrlExpr();
-
-        return query
-                .select(Projections.constructor(
-                        BoardItem.class,
-                        post.id,
-                        preview,
-                        authorIdExpr,
-                        getAuthorName(),
-                        board.category,
-                        post.createdAt,
-                        likedByMe,
-                        bookmarkedByMe,
-                        likeCountExpr,
-                        commentCountExpr,
-                        post.checkCount,
-                        userImageUrlExpr,
-                        contentThumbUrlExpr,
-                        postImageCountExpr(),
-                        Expressions.numberTemplate(Long.class, "NULL")
-                ))
-                .from(post)
-                .join(post.author, user)
-                .join(post.board, board)
-                .where(post.id.in(ids).and(visibleToMe))
-                .fetch();
-    }
 
     @Override
-    public Page<PostListResponse> searchPosts(PostSearchRequest condition, Pageable pageable) {
-        List<PostListResponse> content = query
-                .select(Projections.constructor(PostListResponse.class,
+    public Page<PostListForAdminResponse> searchPostsByAdmin(PostSearchForAdminRequest condition, Pageable pageable) {
+        List<PostListForAdminResponse> content = query
+                .select(Projections.constructor(PostListForAdminResponse.class,
                         post.id,
                         user.name,
                         user.email,
