+++ src/main/java/core/domain/post/service/PostService.java
@@ -1,9 +1,7 @@
 package core.domain.post.service;
 
 import core.domain.board.dto.BoardItem;
-import core.domain.chat.dto.ToggleTranslationRequest;
-import core.domain.post.dto.*;
-import core.domain.post.dto.PostWriteForChatRequest;
+import core.domain.post.dto.comunity.*;
 import core.global.enums.SortOption;
 import core.global.pagination.CursorPageResponse;
 import jakarta.validation.Valid;
