+++ src/main/java/core/domain/user/service/UserAdminService.java
@@ -9,7 +9,7 @@ import core.domain.chat.repository.ChatParticipantRepository;
 import core.domain.chat.repository.ChatRoomRepository;
 import core.domain.comment.dto.RecentCommentDto;
 import core.domain.comment.repository.CommentRepository;
-import core.domain.post.dto.RecentPostDto;
+import core.domain.post.dto.admin.RecentPostDto;
 import core.domain.post.entity.Post;
 import core.domain.post.repository.BlockPostRepository;
 import core.domain.post.repository.PostRepository;
