+++ src/main/java/core/global/service/CrawledDataAdminService.java
@@ -2,7 +2,7 @@ package core.global.service;
 
 import core.domain.board.entity.Board;
 import core.domain.board.repository.BoardRepository;
-import core.domain.post.dto.CrawledDataDto;
+import core.domain.post.dto.crawling.CrawledDataDto;
 import core.domain.post.entity.CrawledData;
 import core.domain.post.entity.Post;
 import core.domain.post.repository.CrawledDataRepository;
